added instructions for django and postgresql
I've added a set of instructions to get started with the django framework and the PostgreSQL database management system. PSQL is scalable and easy to use with a nice user interface called pgAdmin4. If any instruction steps don't work for whatever reason, consult Google first and ask me later - I've managed to install all of this on both Windows and Mac machines.
- Robin